major clean up
