+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/Contents.json
