+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/icon-40@2x.png
