+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/icon-60@2x.png
