+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/icon-83.5@2x.png
