+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/icon-small-50@2x.png
