+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/icon-small@2x.png
