+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/icon-small@3x.png
