+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/notification-icon@3x.png
