+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/notification-icon~ipad.png
