+++ SplURL/AppSupport/AppIcon.xcassets/AppIcon.appiconset/notification-icon~ipad@2x.png
