+++ SplURL/AppSupport/AppIcon.xcassets/Contents.json
