+++ SplURL/AppSupport/Assets.xcassets/Logo.imageset/Contents.json
