+++ SplURL/AppSupport/Assets.xcassets/Logo.imageset/IYqhTsSD.png-medium.png
